+++ CHANGES.md
@@ -23,6 +23,10 @@
   * (rpcli) New option "-k". This option will verify all known keys in
     keys.conf. Verification is done by decrypting a string that was encrypted
     with the original key and checking if the decrypted string is correct.
+  * GameCube: Added partial support for WIA disc images. A copy of the disc
+    header is stored in plaintext in WIA fomrat, so disc header fields and
+    external images are supported. Region code, age ratings, and the internal
+    GCN banner are not supported.
 
 * New systems supported:
   * Windows/DOS: Executables, dynamic link libraries, and other types of